Move all leaf_function tests into test_leaf_function.py
[ghstack-poisoned]
